동일
2. 프로필
 -프로필 수정화면 Fragment 및 구성요소 배치

3. 각 fragment에 상단바 생성

4.main fragment 화면 보완
 -필터링 화면 Fragment 생성
 -지역 listview 팝업 생성
 -연령대 listview 팝업 생성
 -필터링 버튼과 필터링 화면 연결 (X)
 -필터링 화면과 listview 팝업 연결 (X)

5. 모임작성 화면 및 열람화면
 -모임작성 Fragment 생성 : 모임 정보 입력화면 (미완)
 -나이제한 listview 팝업 연결 (연령대listview) (X)
 -모임작성 버튼과 모임작성 Fragment 연결 (X)
 -작성자/이용자별 모임 상세정보 화면 Fragment 생성
 -각 상세정보 화면 구성요소 및 버튼 생성

*필터링과 모임생성에서 선택하는 종류는 원래 하던 방법으로는 계속 실패해서 listview 팝업을 연결하는 식으로 바꿔서 할 것.
*각 버튼과 페이지를 연결하는 것은 페이지뷰 헷갈려서 제대로 알고 다시 할것.
