+++ frontend/app/src/main/res/layout/fragment_create.xml
@@ -78,7 +78,7 @@
         android:layout_marginStart="150dp"
         android:layout_marginTop="30dp">
         <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/et_menu"
+            android:id="@+id/et_create_menu"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="메뉴 입력" />
@@ -90,14 +90,14 @@
         android:layout_marginStart="150dp"
         android:layout_marginTop="130dp">
         <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/et_place"
+            android:id="@+id/et_create_place"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="장소 입력" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <EditText
-        android:id="@+id/et_date"
+        android:id="@+id/et_create_date"
         android:layout_width="120dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="150dp"
@@ -107,7 +107,7 @@
         android:hint="날짜 입력"/>
 
     <EditText
-        android:id="@+id/et_time"
+        android:id="@+id/et_create_time"
         android:layout_width="120dp"
         android:layout_marginStart="280dp"
         android:layout_marginTop="230dp"
@@ -123,7 +123,7 @@
         android:layout_marginTop="350dp">
 
         <RadioButton
-            android:id="@+id/rb_fellowship_all"
+            android:id="@+id/rb_create_fellowship_all"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="150dp"
@@ -132,7 +132,7 @@
             android:textSize="19sp" />
 
         <RadioButton
-            android:id="@+id/rb_fellowship_yes"
+            android:id="@+id/rb_create_fellowship_yes"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
@@ -141,7 +141,7 @@
             android:textSize="19sp" />
 
         <RadioButton
-            android:id="@+id/rb_fellowship_ano"
+            android:id="@+id/rb_create_fellowship_ano"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
@@ -157,7 +157,7 @@
         android:layout_marginStart="150dp"
         android:layout_marginTop="430dp">
         <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/et_num_participants"
+            android:id="@+id/et_create_num_prt"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="인원수 입력" />
