+++ frontend/app/src/main/res/layout/fragment_filtering.xml
@@ -3,6 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".Dialog.FilteringFragment"
     android:layout_width="match_parent"
     android:layout_height="600dp"
     android:background="#FFFFFF"
@@ -56,7 +57,7 @@
         android:orientation="horizontal">
 
         <RadioButton
-            android:id="@+id/rb_d_filtering_all"
+            android:id="@+id/rb_filtering_all"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="15dp"
@@ -66,7 +67,7 @@
             android:textSize="19sp" />
 
         <RadioButton
-            android:id="@+id/rb_d_filtering_yes"
+            android:id="@+id/rb_filtering_yes"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="80dp"
@@ -76,7 +77,7 @@
             android:textSize="19sp" />
 
         <RadioButton
-            android:id="@+id/rb_d_filtering_no"
+            android:id="@+id/rb_filtering_no"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="80dp"
@@ -85,8 +86,6 @@
             android:textColor="@color/black"
             android:textSize="19sp" />
     </RadioGroup>
-
-
 
     <Button
         android:id="@+id/btn_filtering_ok"
